update constant initializers
