+++ src/constants/lib.js
@@ -1,100 +1,33 @@
-const constantModel = require('./model.js');
-const permissionLevels = require('../constants.js').permissionLevels;
-const processMode = require('../environment.js').processMode;
-
-const userConstants = require('../users/constants.js');
-const userModel = require('../users/model.js');
-const userLib = require('../users/lib.js');
-
-const partModel = require('../inventory/model.js');
-const partConstants = require('../inventory/constants.js');
-
-const challengeConstants = require('../challenges/constants.js');
-const challengeModel = require('../challenges/model.js');
-
+const { permissionLevels } = require('../constants.js');
 const { logInfo, logError } = require('../modules/errors.js');
 
-exports.getAmbassadorApplication = () => {
-  return constantModel.getByName('ambassadorApplication');
-};
+const constantsModel = require('./model.js');
+const constantsInitializer = require('./initializer.js');
 
-const createConstantPromise = (
-  constantName, creationFn, data,
-) => {
-  return new Promise((resolve, reject) => {
-    creationFn(data)
-      .then((res) => {
-        if(! res) {
-          reject(new Error('Failed to create constant:', data, res));
-        }
-        if(res._id) {
-          constantModel.createConstant({
-            name: constantName,
-            id: res._id,
-          }).then(() => {
-            logInfo('[+] Created constant:', constantName);
-            resolve([constantName, res]);
-          })
-          .catch(err => {
-            logError('[!] Error resolving createFn err');
-            reject(err);
-          });
-        } else {
-          reject(new Error('Failed to create constant:', data, res.error));
-        }
-      })
-      .catch(err => {
-        logError('[!] Error creating', constantName, ':', err.message);
-        reject(err);
-      });
-  });
-};
+const userModel = require('../users/model.js');
 
-let allPartConstants = {}
-partConstants.allParts.forEach(part => allPartConstants[part.name] = () => 
-  createConstantPromise(
-    part.name,
-    (partData) => partModel.createPart(partData),
-    part,
-  )
-);
+const challengeModel = require('../challenges/model.js');
 
-
-const processModeInitializers = {
-  "stocktracker": allPartConstants,
-  "ambassadorsite": {
-    'adminUser': () => createConstantPromise(
-      'adminUser',
-      (userData) => userLib.createUser(userData),
-      userConstants.adminUserData,
-    ),
-    'ambassadorApplication': () => createConstantPromise(
-      'ambassadorApplication',
-      challengeModel.createChallenge,
-      challengeConstants.ambassadorApplicationData,
-    ),
-  },
-}
 
 /*
  * Setup all constants for Ambassador Server
  *
- * Returns: Thenable
+ * Returns: resultMap (thenable)
  */
 exports.initSiteState = () => {
   let buildfns = []; // functions to check constants and compile create funcs if needed
   let fns = [];      // all create funcs
-  let initializers = processModeInitializers[processMode];
+  let initializer = new constantsInitializer();
 
-  Object.keys(initializers).map(key => {
+  Object.keys(initializer.initializers).map(key => {
     buildfns.push(
-      constantModel.getByName(key)
+      constantsModel.getByName(key)
         .then(res => {
           if(! res) {
             logInfo(
               '[+]', key, 'not found. Creating...'
             )
-            fns.push(initializers[key]());
+            fns.push(initializer.initializers[key]());
           } else {
             logInfo('[+]', key, 'already exists.');
           }
